chore: Update npm dependencies and handle error responses more effectively
This commit updates the npm dependencies in the `package.json` file, adding the `@faker-js/faker` package with version `^8.4.1`. It also includes minor code changes in the `auth.js`, `users.js`, and `action.js` files to handle error responses more effectively and improve the error handling flow in the application.

Please review and merge this commit.